Feat : HP 예외처리추가 (초과방지)

+++ Assets/01.Scrips/9.Item/EquipMentItemData.cs
@@ -16,5 +16,5 @@ public class EquipmentItemData : ItemData
     
    
     public EquipType  equipType;
-    public float valueAmount;
+    public int valueAmount;
 }
